break simulate chart out of node and add "Add Chart" button
